analyzer: further work into codegen IR.
One thing that I've observed while pulling my hair over the design of
this is that I need to do caching, i.e. `codegen_ir()` must behave like
`statement::analyze()`, and it must memoize calls to `_codegen_ir()`,
because otherwise 1) the performance is going to be *really bad*, and 2)
because the codegen itself will have a *very* hard time figuring out
which variables are the same variable and which aren't. (This last point
will still need to be implemented for `id_expression` somehow, but I
think that's ok, since there's an explicit noop instruction for that.)