analyzer: make function creation more lazy.
Fixes an annoying logical race between using a function and being able
to see it. Now a function is visible as soon as its declaration is
visible and arguments resolved. (This is going to be funny once
functions can have arguments, and their types can be specified with
arbitrary expressions. But don't worry, we'll deal with that.)