codegen: generate lower-level C++ with "registers" for variables.
C++ is awkwardly too high level for some of the rules of Vapor I want to
implement, hence this change. The resulting generated code is horrible
and in general completely (or almost completely?) unoptimizable... I
hope I will be able to make that better soon-ish. There are two obvious
next steps here:

1) Actually implement object lifetimes. `materialization_instruction`
and the rest of the instructions do well with constructing an object
(I'm 99% sure that once I start generating something else than just C++
this will get far more complicated, but shush for now), but there's
literally no place that destroys objects. Need to do this at the end of
scopes, which will be the first step to actually do the correct thing
here.

2) Implement variable usage analysis, and then automatically move
variables on their last use. This is deeply necessary for sanity (both
of the programmer and of the generated C++).

Then... outside of this branch, probably the next step would be
user-defined types and tuples, and function arguments.