So solving the recursive case without this hint is hard.
Basically it requires the compiler to first notice that the type of the
last return expression depends... on itself. This alone will require
building an actual dependency graph - I wanted to do this either way,
but this is just another thing that requires it - and noticing that
fact; then, it boils down to solving the following:

x - return type of `fibonacci`

(1) x = int (for all the constant cases)
(2) x = typeof(x{} + x{})

From these two, it should basically solve for x by substitution (this is
the moment when I start writing a sort of a type unifier, which will
also be extremely helpful later on for... virtually everything):

x = typeof(int{} + int{}) // substitute `x` from (1) into rhs of (2)
x = int

This looks like it could just substitue `int` for `x` on both sides, but
I don't think that's the case, since there is a possibility of these
types not actually being one type.

(Isn't the reasoning gibberish at that point though? If the types are
obviously different - `return 1;` vs `return 1.0;` - a sensible decision
*could* be made, though I'm not sure if converting either is actually
sensible and whether this shouldn't just be a hard error. By
substituting both sides in this case, the result is `int = double`,
which is a contradiction, on which we can error. In the fibonacci case,
it'd become `int = typeof(int{} + int{})`, or `int = int`, which is
sensible, which could mean that the whole type will typecheck, so the
one concrete return type is correct.)

(Obviously there's a lot of tradeoffs to make here. I believe that one
possible resolution of a case where the set of equations is

x = int
x = double

is to find a common typeclass - in this case Numeric or something,
whatever is its name in Vapor - and return a type-erased variable of
that typeclass, which'd *possibly* be sensible. Not entirely sure though
which behavior is preferred, and not entirely sure how to decide which
typeclass to select if both types are members of multiple typeclasses.
Possibly some logic to check which one is the most specific to them -
say, defines the most members of both - and select that one.)

(Either way, both cases are far away. Currently, I'll need to add a
dependency graph and error out in this non-trivial recursive case, or
incompatible types case, and just require a type specifier for the
function with such shenanigans.)