analyzer: make things work for recursive functions.
There's a few significant changes here. First, there's guards against
infinite recursion in some cases, to avoid stack overflows. The code
may look somewhat weird, for example in generation of IR for types and
so on, but it works pretty great.

The other important change is caching a copy of blocks. This is to
allow simplification of recursive functions, because otherwise the
internal expressions would attempt to wait for the block to be
simplified, which in turn waits for them, resulting in no progress
whatsoever. This will result in somewhat more work in some degenerate
cases of compile-time evaluation, but whatever.

One significant problem that currently exists - which is reflected by
the change to `main.cpp` - is that simplification of blocks is done in a
somewhat silly way of doing everything parallelly, which means that
early returns aren't currently understood. Changes to fix this are to
follow, but I don't have them ready yet, and I want to commit all the
progress that I currently have.