analyzer: attempts at getting simplification working with structs.
I NEED TESTS.

No, seriously, this is the moment when I stop the world and start
actually writing tests for the analyzer, because otherwise I'll die. I
need them both for the semantic part and for the simplification part.

These will be tests that start with parsing and end with checking some
conditions. It's not yet time to work on getting actual "here's input,
do your thing" tests for the full compilation process going.