analyzer: remove the class hierarchy of variables.
A while ago it became apparent to me that this concept is not truly
needed, and is actually harmful for how the code is structured,
including some ownership issues that I've had. This commit removes the
class variable, syntactically, from the project.

Do note that this is not a finished refactor and as such this commit is
not meant to be working. It's probably here to stay, to track how much
work went into this, but it was never meant to be a functioning commit.

Now, the code is valid syntactically, meaning it compiles (and even
links), but there's a lot of testcases that either downright crash (at
least with ASAN) or time out in some way. There's going to be a big
effort in fixing those tests, and then an equally big effort to also
fill in the missing pieces that I deemed too complex to fill in while
removing `variable` from the code. For some of them I have proper ideas
for how to handle them; for others not so much, and they'll need some
architectural design to work well, but I am sure that I will eventually
find a way to solve those problems, and I'm confident that this change
fixes a deep design problem that would not have been maintainable in the
long run.