analyzer: sanitize the use of member expression.
It was sometimes used where member_access_expression was more
appropriate. This would make it impossible to codegen actual member
expressions to pass around.

Also added a hook for IR invalidation. It is particularly important in
member access, since the same generic expression object will be used for
all calls of the replacement copy constructor, which means that whenever
the base expression changes - we need to regenerate the IR.