analyzer: support a subset of recursive calls for constant folding.
This supports recursion that either has all the arguments be constants
(but at least somewhat *different* from the previous call frames), or
ones where the constness of the arguments changes across the calls. The
one thing that is not yet supported is recursive calls where multiple
calls with some specific arguments are non-const, but are *different*
variables, since we don't have a way to distinguish between different
non-constant expressions (across clones). This is incoming soon-ish but
is not as big of priority as the change here.

Before I merge this, though, I really need to get result caching for
all-constant-arguments calls going, because the current running time of
simplification of A(3, 4) is absurd (7.5 minutes on my machine,
currently).