analyzer: make AAST nodes constructible without AST nodes.
This will enable making the analyzer unit testable by not having to
parse anything in order to build various expressions. Some tests are to
follow, especially since the next feature I want to work on is
typeclasses, which will completely revamp how overload resolution works,
in pretty much all cases. (This'll be fun... but only with tests.)