analyzer; begin the work on properly supporting modules.
The first step is to make a module a statement. This allows doing some
of the work necessary for them to work without having to spend time
wondering how they will look like when they are actual expressions.

Currently breaks tests, because the __entry_call_thunk function is not
generated, because no entry point is found. This will be addressed in
the next commit, where I actually implement the special casing for
module `main`. This will be a special module that can appear in any
file (relevant now, even though I still have to implement the filename
checking... eh, work work work), and `main.entry` is always going to be
the function to be called as the entry point.

Depending on how the codegen for typeclasses actually looks like, this
may allow me to eventually fully get rid of the __entry_call_thunk, but
that's currently a non-goal.