general: start using CMake in anticipation of using protobuf.
I'm doing multiple builds in multiple configurations, with multiple
stdlibs, so the standard packages for protobuf aren't really what I
want. I'll probably implement a switch to use the system protobuf
instead of the locally built one one day, but right now I don't need
that. For the time being, protobuf is going to be statically linked into
libvprc.so.

Protobuf is going to be used for the module interface files; to what
extend, we will see. On one hand, it seems that the entire module will
need to be saved in there, because arbitrary functions are meant to be
callable at compile time. On the other hand, we're going to have to stop
at some point. I'll probabaly end up having an object file for the
module, and interface file for just the interface, and an implementation
file with things callable at compile time.

There will need to exist some ways to limit what lands in the
implementation file, because otherwise insanity is going to ensue.
Possibly a qualificator for what can be callable at compile time, but
that seems as bad as C++'s `constexpr`. Certainly all templates that are
exported need to land in the interface file, but it's the other things
that I am very murky about.

CMake is also not the end goal; I want to one day use despayre instead.
We'll see how that goes; there's a lot of functionality missing from
despayre for that to happen right now.