analyzer: first steps of importing modules.
The current code can find compiled or not modules in the specified
locations, detect whether it needs to recompile a compiled module
interface file that's older than its source, and then asserts in all
cases.

Next step is making the compiler generate an interface file. I need to
figure out where exactly to generate them; I'm currently of the mind
that is should be a part of the analyzer, since the internal IR might
not have all the necessary information, but we'll see.

Then, I'll implement loading a module from an interface file. At the
very end will come make-like invocation that compiles all the
dependencies that aren't yet compiled (as validating this REQUIRES the
previous steps, while the previous steps can be validated without this).