analyzer: introduce the concept of associated entities.
For now they'll serve basically just to import the invented entities
created for overload set types. In the future, I imagine that this is
how typeclass instances will be imported.

Also import overload sets in the form that they currently have. Still
need to learn to export them, and then import the signatures.