analyzer: fix a bunch of lifetime issues.
Had some really bad lifetime management in some places, and some
unique_ptr destructors were really ruining the day. It seems fixed now.
Also `import_expression` is now, to an extend, and actual expression.

The next step is to actually write code for unresolved types to resolve
themselves. This is going to be fun.