analyzer: make typeclass preanalysis more like struct's.
The bulk of the datastructure creation now lies on the typeclass type
side, and not on the side of the expression, which (1) is more in line
with what I did before for structures, and (2) actually makes more
sense, separating concerns nicely and making it easier to mock the type
if I want to do some sort of unit tests on the typeclass_literal class.

Also some tiny grammar nazing (possibly incorrect, but I like it more
the new way).