analyzer: progress in instance creation.
A missing assert that I've added back now notified me that I will
actually need to fix the member/nonmemberness of functions (especially
of overload_set functions).

This commit has *more* than strictly necessary, i.e. setup of the scope
of the instance, which will be necessary when actually turning typeclass
instances into types (...which *may* happen as a part of a separate
branch, because doing all that here *may* end up being a gigantic
change).