analyzer: some further work on typeclasses.
I realized that I can't *really* skip properly analyzing definitions in
the typeclass body for the MVP of this feature; it'd be incredibly
weird, and default implementations of functions defined inline would
just not work (and while that isn't a deal breaker for the built-in
usage that I have in mind right now, it'd be *hella* annoying).

Therefore, the next step on this crazy journey is to generate
archetypes that behave like (but aren't) types, to enable actually
analyzing all the code (and not having to deal with token soup and duck
typing).

At first I'll generate an archetype for `type`. Later, I'll add a way
for typeclasses (that constrain the type) to modify the archetype, when
constraints are present). There will eventually be other kinds of
archetypes, for types of higher kinds, but since those are only
somewhere on the horizon - so are their archetypes.