analyzer: add typeclass and instance serialization.
Visual inspection of deserialized vprm for the exports test suggests it
is correct. Of course that's no proof, but since there's no
deserialization code written yet, that's the best we can do right this
second. Next step is to add some tests where I actually export and
import things in interesting ways and try to make them work. I'm not
sure if what I did here is sufficient for the import side, but in the
worst case I can always reshuffle everything in here right now.