analyzer: implement support for importing typeclasses.
One big missing thing is importing default implementations of typeclass
member functions, but that will come at a later date when I figure out
how the heck to serialize function definitions.

This is the first commit that touches in #52; it is by no means a proper
implementation of that, since it mostly only _reads_ the AST location
info from the module files; it doesn't really _write_ them, which is
something that will require a more proper pass over everything we
export.

This also somewhat improves dependencies for proto files; it's slightly
overeager there, because all proto files are re-compiled when any of
them changes.

To make testcases more interesting, integer division has been
implemented.