analyzer: implement importing typeclass instances.
This change is accompanied by a rather big change in how modules work:
the concept of "associated entities" in the serialized module file is
now gone; it still exists at the stage where the file is generated, but
not when it is loaded. (This is so far very heavily stubbed; see all the
FIXMEs.)

Also, there's now entities that can be exported, but not have their name
exported; this was necessary for typeclass instance types, but will in
the future make many more things possible. Overload sets are now also
exported in this manner. An exported entity is only findable by name in
its scope if the appropriate field in the serialized entity is set.

More name propagation was necessary for this; what is currently in is
definitely not complete, but it's rather hard to get this fully
completed, so I guess this will be an ongoing effort for a while.

As a consequence of all this, unresolved user defined references are now
looked up much more sanely, without recursive scope lookups like they
used to (because they can now find entities that _are not_ added to
lexical scopes).

TODO right now: redo how functions are added to IR to be generated. The
current way of forcing generation of everything that is used doesn't
seem to mesh too well with functions with external definitions we now
have, so I'll probably rewrite that entire thing just like I rewrote a
lot of modules in this commit ;)