analyzer, codegen: finish instance imports.
Two big changes included. One, functions are no longer generated
whenever a call instruction to one is emitted; that was troublesome to
make work with imported member functions of instances. This was a hack
that was created "back in the days" for "intrinsic" functions, because
they were emitted as actual functions, generated on demand. Now they
behave like proper intrinsics, inserting their instructions into the
instruction stream of a call expression, so that hack could be removed.
Intrinsics now being proper is the second big change.